feat(tracker): announce and scrape, with access as an entitlement
The BitTorrent endpoints ported, plus the mem store that completes the triple.

THE ACCESS CHECK IS THE INTERESTING PART. The host asked
`!u.TrackerAccess || u.RoleLevel() <= RoleDisabled` against a user row it owned.
The plugin cannot read that row, so the check splits into two questions the Gate
struct keeps separate on purpose: Entitled is the tracker's own decision, now
core.Entitlements.Has(ctx, userID, "tracker.access"), and Active is whether the
account still exists and is not banned, which is the host's business.

Both fail CLOSED when the service is absent, and that is not defensive
boilerplate. A host with no entitlements service has not decided that everyone may
use the tracker -- it has not wired the thing that decides. Answering yes on its
behalf would open a private tracker to anyone holding a passkey. There is a test
for it, because the failure would be silent.

Two protocol details worth keeping visible now they live somewhere new. A failure
answers HTTP 200 carrying a bencoded failure reason, because that is what a client
DISPLAYS -- a 403 shows up as an unexplained "tracker error", so the status code is
the wrong channel. And the snatch increments on the TRANSITION only
(delta.Done && !prev.Completed), which is what stops a client that keeps
announcing after finishing from incrementing the count every few minutes for a
week.

The peer_id/member check survives the move: one peer_id belongs to one member.
The passkey says who is announcing, but without this a second person could
announce the same peer_id and have their traffic credited to the first -- which is
the whole "copy a friend's .torrent" path.

Peer.UserID widens from int to int64 to match core.User.ID and the store. Free:
the field is JSON on the wire where a number is a number, and there were no peers
in Redis to reinterpret.

I WROTE A VANITY ASSERTION AND CAUGHT IT. The totals test meant to prove that a
stale complete peer is not counted as seeding, and it did two things wrong: it
moved the clock back and forward again before writing anything, so every row was
stamped "now" and the stale case never existed, and the check was a t.Logf, which
is why it passed regardless. Rewritten to set the clock before each write, and
mutation-checked -- counting seeding from left_bytes alone now fails it. That
matters because reporting a stale peer as seeding tells a member they are seeding
a torrent their client stopped days ago.

+++ tracker/peers.go
@@ -23,7 +23,10 @@ const PeerTTL = 45 * time.Minute
 // deltas — the announce handler diffs them against the previous snapshot to
 // compute how much to add to the per-user bytes.
 type Peer struct {
-	UserID     int    `json:"u"`
+	// int64, widened from the host's int so it matches core.User.ID and the
+	// store. Free to change: the field is JSON on the wire, where a number is a
+	// number, and there were no peers in Redis to reinterpret.
+	UserID     int64  `json:"u"`
 	PeerID     string `json:"p"`
 	IP         string `json:"i"`
 	Port       int    `json:"o"`
